requirements.txt
# requirements.txt
opencv-python==4.9.0.80
mediapipe==0.10.11
numpy==1.26.4
python-dotenv==1.0.1
black==25.1.0